# Optimize bdf