@jepler 2020-12-09
We also benchmarked this with a substantially larger font (latin + kana) and the improvement was marginal.  Rather than optimize bdf at the expense of readability, we've implemented a pcf (binary font) reader.